third_party/blink/web_tests/external/wpt/fetch/orb/tentative/img-mime-types-coverage.tentative.sub-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] ORB should block the response if Content-Type is: 'application/dash+xml'.  
  assert_unreached: Unexpected load event Reached unreachable code
[FAIL] ORB should block the response if Content-Type is: 'application/vnd.apple.mpegurl'.  
  assert_unreached: Unexpected load event Reached unreachable code
[FAIL] ORB should block the response if Content-Type is: 'application/x-protobuffer'.  
  assert_unreached: Unexpected load event Reached unreachable code
[FAIL] ORB should block the response if Content-Type is: 'audio/mpegurl'.  
  assert_unreached: Unexpected load event Reached unreachable code
[FAIL] ORB should block the response if Content-Type is: 'text/vtt'.  
  assert_unreached: Unexpected load event Reached unreachable code
Harness: the test ran to completion.
